Rename pathVisualizer.ts to index.ts
